Fix MSVC warnings in client.c, server.c, links.c, main.c
- Change LINK_DEFS_COUNT from extern const size_t to #define so it can
  be used as a compile-time array size (fixes VLA error in main.c)
- Change loop variables to int8_t where used as player/client IDs
- Widen send_player_action and send_discards_request_new_cards return
  types from int8_t to int to match send_all_tcp
- Change coin_array_size field from size_t to uint32_t to match
  pcg32_boundedrand_r parameter type
- Replace double literals in layout arithmetic with integer equivalents
  (e.g. * 0.1 -> / 10, / 1.5 -> * 2 / 3, * 0.82 -> * 82 / 100)
- Refactor selected_nick ternary assignments to if/else to avoid
  int-to-int8_t narrowing from ternary promotion
- Add boundary cast for strlen result assigned to uint16_t wire length

+++ src/client.h
@@ -64,7 +64,7 @@ typedef struct {
   ma_engine_config engineConfig;
   Sound_t *sounds;
   Sound_t *coin_hit_sounds;
-  size_t coin_array_size;
+  uint32_t coin_array_size;
 } SoundContext_t;
 
 extern const GameChoice_t game_choices[];
@@ -75,13 +75,13 @@ bool get_socket_context_and_run_client(PlayerConfig_t *player_config, const CliA
                                        const bool test_mode, LinkWidget_t **links,
                                        SocketContext_t *out_socket_context);
 
-int8_t send_player_action(ClientState_t *client_state, TCPsocket sock, uint8_t action,
-                          uint32_t amount);
+int send_player_action(ClientState_t *client_state, TCPsocket sock, uint8_t action,
+                       uint32_t amount);
 
 void do_sdl_cleanup(SdlContext_t *sdl_context);
 
-int8_t send_discards_request_new_cards(TCPsocket sock, const uint8_t *discard_indices,
-                                       uint8_t count);
+int send_discards_request_new_cards(TCPsocket sock, const uint8_t *discard_indices,
+                                    uint8_t count);
 
 void layout_cards(CardContext_t card_context[MAX_PLAYERS][MAX_HAND_SIZE], Player_t *players_array,
                   const SDL_Point *player_pos);
